reconcile: address review change requests
